Create production image build workflow
Adds a new workflow to build the production image for pushing to the container registry. It uses github actions to specify the build job and sets up a sequence of steps that include setting environment variables, checking out source code, docker metadata, and building and pushing the image to the specified repository using the Docker build-push action. This workflow will only run on the main branch and is triggered by pushes to the main branch.